pyproject.toml
[project]
name = "weather-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
